+++ src/commands/controller.ts
@@ -16,23 +16,25 @@ class {className} extends BaseController
 
 `;
 
-let value: string = '';
 let name: string = '';
 let filename: string = '';
 let body: string = '';
 
 const controller = async (vscode: any, fs: any, path: any) => {
-  value = await vscode.window.showInputBox({
+  name = await vscode.window.showInputBox({
     prompt: 'Controller class name',
-    placeHolder: 'Controller class name',
+    placeHolder: 'E.g. UserController, AuthController...',
     validateInput: (text: string) => {
-      if (!/^[a-zA-Z]\w+$/.test(text)) {
-        return 'Invalid format!';
+      if (!/^[A-Z][A-Za-z]{2,}$/.test(text)) {
+        return 'Invalid format! Class names MUST be declared in StudlyCaps / PascalCase (psr-1)';
       }
     },
   });
 
-  name = value.charAt(0).toUpperCase() + value.replace(/\s/g, '').slice(1);
+  if (name.length === 0) {
+    return;
+  }
+
   filename = '/app/Controllers/' + name + '.php';
 
   body = content.replace('{className}', name);