+++ src/commands/resource.ts
@@ -50,34 +50,36 @@ class {className} extends ResourceController
 
 `;
 
-let value: string = '';
 let model: string = '';
 let name: string = '';
 let filename: string = '';
 let body: string = '';
 
 const helper = async (vscode: any, fs: any, path: any) => {
-  value = await vscode.window.showInputBox({
+  name = await vscode.window.showInputBox({
     prompt: 'Resource class name',
-    placeHolder: 'Resource class name',
+    placeHolder: 'E.g. UserController, AuthController...',
     validateInput: (text: string) => {
-      if (!/^[a-zA-Z]\w+$/.test(text)) {
-        return 'Invalid format!';
+      if (!/^[A-Z][A-Za-z]{2,}$/.test(text)) {
+        return 'Invalid format! Class names MUST be declared in StudlyCaps / PascalCase (psr-1)';
       }
     },
   });
 
   model = await vscode.window.showInputBox({
     prompt: 'Model class name',
-    placeHolder: 'Model class name',
+    placeHolder: 'E.g. UserModel, GroupModel...',
     validateInput: (text: string) => {
-      if (!/^[a-zA-Z]\w+$/.test(text)) {
-        return 'Invalid format!';
+      if (!/^[A-Z][A-Za-z]{2,}$/.test(text)) {
+        return 'Invalid format! Class names MUST be declared in StudlyCaps / PascalCase (psr-1)';
       }
     },
   });
 
-  name = value.replace(/\s/g, '_').toLowerCase();
+  if (name.length === 0) {
+    return;
+  }
+
   filename = '/app/Controllers/' + name + '.php';
 
   body = content.replace('{className}', name).replace('{modelName}', model);