+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2023-08-30
+
+### Added
+
+- Add Migrate Commands
+
+### Fixed
+
+- Fix formatting and validation of file names
+
 ## [1.1.0] - 2023-08-28
 
 ### Added
@@ -27,7 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/releases/tag/v1.0.0